Changed the attendees service file into the EventService and created EventServiceTest
Deleted AttendeesService and AttendeesServiceTest
